Extract override foreign key name and translations types
MetadataEntityOverridableForeignKeyName moves next to
MetadataManyToOneJoinColumn under flat-entity/types, and
MetadataEntityOverridesTranslations gets its own file, so each generic reads
as one type.
